+++ core/src/main/scala/japgolly/scalajs/react/ReactComponentB.scala
@@ -1,121 +1,158 @@
 package japgolly.scalajs.react
 
-import scala.scalajs.js._
+import scala.scalajs.js.{Any => JAny, _}
 import Internal._
 
 object ReactComponentB {
-  def apply[Props](name: String) = new ReactComponentB[Props](name)
-}
+  def apply[Props](name: String) = new P[Props](name)
 
-final class ReactComponentB[Props](name: String) {
+  // ===================================================================================================================
+  final class P[Props] private[ReactComponentB](name: String) {
 
-  def getInitialState[State](f: Props => State) = new B2[State](f)
-  def initialState[State](s: => State)          = getInitialState(_ => s)
-  def stateless                                 = initialState(())
+    @deprecated("getInitialState() has been renamed to initialStateP() and will be removed in 0.6.0.", "0.5.0")
+    def getInitialState[State](f: Props => State) = initialStateP(f)
 
-  def render(f: Props                  => VDom) = stateless.render((p,_) => f(p))
-  def render(f: (Props, PropsChildren) => VDom) = stateless.render((p,c,_) => f(p,c))
+    def initialStateP[State](f: Props => State) = new PS[Props, State](name, f)
+    def initialState[State](s: => State)        = initialStateP(_ => s)
+    def stateless                               = initialState(())
 
-  // -----------------------------------------------------------------------------------------------
-  final class B2[State] private[ReactComponentB](getInitialState: Props => State) {
-    type ScopeB = BackendScope[Props, State]
+    def render(f: Props                  => VDom) = stateless.render((p,_) => f(p))
+    def render(f: (Props, PropsChildren) => VDom) = stateless.render((p,c,_) => f(p,c))
+  }
 
-    def backend[Backend](f: ScopeB => Backend) = new B3[Backend](f)
-    def noBackend                              = new B3[Unit](_ => ())
+  // ===================================================================================================================
+  final class PS[Props, State] private[ReactComponentB](name: String, initF: Props => State) {
+
+    def backend[Backend](f: BackendScope[Props, State] => Backend) = new PSB[Props, State, Backend](name, initF, f)
+    def noBackend                                                  = backend(_ => ())
 
     def render(f: (Props, State)                                       => VDom) = noBackend.render((p,s,_) => f(p,s))
     def render(f: (Props, PropsChildren, State)                        => VDom) = noBackend.render((p,c,s,_) => f(p,c,s))
     def render(f: ComponentScopeU[Props, State, Unit]                  => VDom) = noBackend.render(f)
     def renderS(f: (ComponentScopeU[Props, State, Unit], Props, State) => VDom) = noBackend.renderS(f)
+  }
 
-    // ---------------------------------------------------------------------------------------------
-    final class B3[Backend] private[ReactComponentB](backend: ScopeB => Backend) {
-      type ScopeU  = ComponentScopeU [Props, State, Backend]
-      type ScopeM  = ComponentScopeM [Props, State, Backend]
-      type ScopeWU = ComponentScopeWU[Props, State, Backend]
+  // ===================================================================================================================
+  final class PSB[P, S, B] private[ReactComponentB](name: String, initF: P => S, backF: BackendScope[P, S] => B) {
 
-      type CCP  = CompCtorP [Props, State, Backend]
-      type CCOP = CompCtorOP[Props, State, Backend]
-      type CCNP = CompCtorNP[Props, State, Backend]
+    def render(f: ComponentScopeU[P, S, B] => VDom): ReactComponentB[P, S, B] =
+      new ReactComponentB(name, initF, backF, f,
+        LifeCycle(undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined))
 
-      def renderS(f: (ScopeU, Props, State)                => VDom): B4[CCP] = render(T => f(T, T.props, T.state))
-      def render(f: (Props, State, Backend)                => VDom): B4[CCP] = render(s => f(s.props, s.state, s.backend))
-      def render(f: (Props, PropsChildren, State, Backend) => VDom): B4[CCP] = render(s => f(s.props, s.propsChildren, s.state, s.backend))
-      def render(f: ScopeU                                 => VDom): B4[CCP] =
-        B4[CCP](f, new CompCtorP(_, None)
-          , undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined)
+    def render(f: (P, S, B) => VDom): ReactComponentB[P, S, B] =
+      render(s => f(s.props, s.state, s.backend))
 
-      // -------------------------------------------------------------------------------------------
-      case class B4[C] private[ReactComponentB](
-          __render: ScopeU => VDom,
-          __c: ComponentConstructor[Props, State, Backend] => C
-          , getDefaultProps           : UndefOr[() => Props]
-          , componentWillMount        : UndefOr[ScopeU                  => Unit]
-          , componentDidMount         : UndefOr[ScopeM                  => Unit]
-          , componentWillUnmount      : UndefOr[ScopeM                  => Unit]
-          , componentWillUpdate       : UndefOr[(ScopeWU, Props, State) => Unit]
-          , componentDidUpdate        : UndefOr[(ScopeM, Props, State)  => Unit]
-          , componentWillReceiveProps : UndefOr[(ScopeM, Props)         => Unit]
-          , shouldComponentUpdate     : UndefOr[(ScopeM, Props, State)  => Boolean]
-          ) {
+    def render(f: (P, PropsChildren, S, B) => VDom): ReactComponentB[P, S, B] =
+      render(s => f(s.props, s.propsChildren, s.state, s.backend))
 
-        def getDefaultProps(f: => Props): B4[C]    = copy(getDefaultProps = () => f)
-        def propsDefault   (f: => Props): B4[CCOP] = copy(__c = new CompCtorOP(_, None, () => f))
-        def propsAlways    (f: => Props): B4[CCNP] = copy(__c = new CompCtorNP(_, None, () => f))
-        // def propsAlways(f: => Props): B4[CCNP] = getDefaultProps(f).copy(__c = new CompCtorNP(_))
+    def renderS(f: (ComponentScopeU[P, S, B], P, S) => VDom): ReactComponentB[P, S, B] =
+      render(T => f(T, T.props, T.state))
+  }
 
-        def componentWillMount       (f: ScopeU                  => Unit   ): B4[C] = copy(componentWillMount        = fcUnit(componentWillMount       , f))
-        def componentDidMount        (f: ScopeM                  => Unit   ): B4[C] = copy(componentDidMount         = fcUnit(componentDidMount        , f))
-        def componentWillUnmount     (f: ScopeM                  => Unit   ): B4[C] = copy(componentWillUnmount      = fcUnit(componentWillUnmount     , f))
-        def componentWillUpdate      (f: (ScopeWU, Props, State) => Unit   ): B4[C] = copy(componentWillUpdate       = fcUnit(componentWillUpdate      , f))
-        def componentDidUpdate       (f: (ScopeM, Props, State)  => Unit   ): B4[C] = copy(componentDidUpdate        = fcUnit(componentDidUpdate       , f))
-        def componentWillReceiveProps(f: (ScopeM, Props)         => Unit   ): B4[C] = copy(componentWillReceiveProps = fcUnit(componentWillReceiveProps, f))
-        def shouldComponentUpdate    (f: (ScopeM, Props, State)  => Boolean): B4[C] = copy(shouldComponentUpdate     = fcEither(shouldComponentUpdate  , f))
+  // ===================================================================================================================
+  private[react] case class LifeCycle[P,S,B](
+    getDefaultProps           : UndefOr[()                                => P],
+    componentWillMount        : UndefOr[ComponentScopeU[P, S, B]          => Unit],
+    componentDidMount         : UndefOr[ComponentScopeM[P, S, B]          => Unit],
+    componentWillUnmount      : UndefOr[ComponentScopeM[P, S, B]          => Unit],
+    componentWillUpdate       : UndefOr[(ComponentScopeWU[P, S, B], P, S) => Unit],
+    componentDidUpdate        : UndefOr[(ComponentScopeM[P, S, B], P, S)  => Unit],
+    componentWillReceiveProps : UndefOr[(ComponentScopeM[P, S, B], P)     => Unit],
+    shouldComponentUpdate     : UndefOr[(ComponentScopeM[P, S, B], P, S)  => Boolean])
+}
 
-        def buildSpec = {
-          val spec = Dynamic.literal(
-              "displayName" -> name,
-              "backend" -> 0,
-              "render" -> (__render: ThisFunction)
-            )
+import ReactComponentB.LifeCycle
 
-          @inline def setFnPS[T, R](fn: UndefOr[(T, Props, State) => R], name: String): Unit =
-            fn.foreach { f =>
-              val g = (t: T, p: WrapObj[Props], s: WrapObj[State]) => f(t, p.v, s.v)
-              spec.updateDynamic(name)(g: ThisFunction)
-            }
+final class ReactComponentB[P, S, B](name: String,
+                                     initF: P => S,
+                                     backF: BackendScope[P, S] => B,
+                                     rendF: ComponentScopeU[P, S, B] => VDom,
+                                     lc: LifeCycle[P, S, B]) {
 
-          val componentWillMount2 = (t: ScopeU) => {
-            val scopeB = t.asInstanceOf[ScopeB]
-            t.asInstanceOf[Dynamic].updateDynamic("backend")(backend(scopeB).asInstanceOf[Any])
-            componentWillMount.foreach(g => g(t))
-          }
-          spec.updateDynamic("componentWillMount")(componentWillMount2: ThisFunction)
+  @inline private implicit def lcmod(a: LifeCycle[P, S, B]): ReactComponentB[P, S, B] =
+    new ReactComponentB(name, initF, backF, rendF, a)
 
-          val initStateFn: ScopeU => WrapObj[State] = scope => WrapObj(getInitialState(scope.props))
-          spec.updateDynamic("getInitialState")(initStateFn: ThisFunction)
+  def getDefaultProps(p: => P): ReactComponentB[P, S, B] =
+    lc.copy(getDefaultProps = () => p)
 
-          getDefaultProps.foreach(f => spec.updateDynamic("getDefaultProps")(f: Function))
-          componentWillUnmount.foreach(f => spec.updateDynamic("componentWillUnmount")(f: ThisFunction))
-          componentDidMount.foreach(f => spec.updateDynamic("componentDidMount")(f: ThisFunction))
+  def componentWillMount(f: ComponentScopeU[P, S, B] => Unit): ReactComponentB[P, S, B] =
+    lc.copy(componentWillMount = fcUnit(lc.componentWillMount, f))
 
-          setFnPS(componentWillUpdate, "componentWillUpdate")
-          setFnPS(componentDidUpdate, "componentDidUpdate")
-          setFnPS(shouldComponentUpdate, "shouldComponentUpdate")
+  def componentDidMount(f: ComponentScopeM[P, S, B] => Unit): ReactComponentB[P, S, B] =
+    lc.copy(componentDidMount = fcUnit(lc.componentDidMount, f))
 
-          componentWillReceiveProps.foreach { f =>
-            val g = (t: ScopeM, p: WrapObj[Props]) => f(t, p.v)
-            spec.updateDynamic("componentWillReceiveProps")(g: ThisFunction)
-          }
+  def componentWillUnmount(f: ComponentScopeM[P, S, B] => Unit): ReactComponentB[P, S, B] =
+    lc.copy(componentWillUnmount = fcUnit(lc.componentWillUnmount, f))
 
-          spec.asInstanceOf[ComponentSpec[Props, State, Backend]]
+  def componentWillUpdate(f: (ComponentScopeWU[P, S, B], P, S) => Unit): ReactComponentB[P, S, B] =
+    lc.copy(componentWillUpdate = fcUnit(lc.componentWillUpdate, f))
+
+  def componentDidUpdate(f: (ComponentScopeM[P, S, B], P, S) => Unit): ReactComponentB[P, S, B] =
+    lc.copy(componentDidUpdate = fcUnit(lc.componentDidUpdate, f))
+
+  def componentWillReceiveProps(f: (ComponentScopeM[P, S, B], P) => Unit): ReactComponentB[P, S, B] =
+    lc.copy(componentWillReceiveProps = fcUnit(lc.componentWillReceiveProps, f))
+
+  def shouldComponentUpdate(f: (ComponentScopeM[P, S, B], P, S) => Boolean): ReactComponentB[P, S, B] =
+    lc.copy(shouldComponentUpdate = fcEither(lc.shouldComponentUpdate, f))
+
+  def propsRequired         = new Builder(new ReactComponentC.ReqProps(_, None))
+  def propsDefault(p: => P) = new Builder(new ReactComponentC.DefaultProps(_, None, () => p))
+  def propsConst(p: => P)   = new Builder(new ReactComponentC.ConstProps(_, None, () => p))
+
+  @deprecated("propsAlways() has been renamed to propsConst() and will be removed in 0.6.0.", "0.5.0")
+  def propsAlways(p: => P)  = propsConst(p)
+
+  // Convenience methods
+
+  @inline def buildSpec = propsRequired.buildSpec
+  @inline def create    = propsRequired.create
+
+  @inline def propsUnit(implicit ev: Unit =:= P) = propsConst(ev(()))
+  @inline def createU(implicit ev: Unit =:= P)   = propsUnit.create
+
+  // ===================================================================================================================
+  final class Builder[C] private[ReactComponentB](cc: ComponentConstructor[P, S, B] => C) {
+
+    def buildSpec: ComponentSpec[P, S, B] = {
+      val spec = Dynamic.literal(
+        "displayName" -> name,
+        "backend" -> 0,
+        "render" -> (rendF: ThisFunction)
+      )
+
+      @inline def setFnPS[T, R](fn: UndefOr[(T, P, S) => R], name: String): Unit =
+        fn.foreach { f =>
+          val g = (t: T, p: WrapObj[P], s: WrapObj[S]) => f(t, p.v, s.v)
+          spec.updateDynamic(name)(g: ThisFunction)
         }
 
-        def create = __c(React.createClass(buildSpec))
-
-        /** When the Props type is Unit, this shortcut returns a constructor that won't ask for a props value. */
-        def createU(implicit ev: Unit =:= Props) = propsAlways(ev(())).create
+      val componentWillMount2 = (t: ComponentScopeU[P, S, B]) => {
+        val scopeB = t.asInstanceOf[BackendScope[P, S]]
+        t.asInstanceOf[Dynamic].updateDynamic("backend")(backF(scopeB).asInstanceOf[JAny])
+        lc.componentWillMount.foreach(g => g(t))
       }
+      spec.updateDynamic("componentWillMount")(componentWillMount2: ThisFunction)
+
+      val initStateFn: ComponentScopeU[P, S, B] => WrapObj[S] = scope => WrapObj(initF(scope.props))
+      spec.updateDynamic("getInitialState")(initStateFn: ThisFunction)
+
+      lc.getDefaultProps.foreach(f => spec.updateDynamic("getDefaultProps")(f: Function))
+      lc.componentWillUnmount.foreach(f => spec.updateDynamic("componentWillUnmount")(f: ThisFunction))
+      lc.componentDidMount.foreach(f => spec.updateDynamic("componentDidMount")(f: ThisFunction))
+
+      setFnPS(lc.componentWillUpdate, "componentWillUpdate")
+      setFnPS(lc.componentDidUpdate, "componentDidUpdate")
+      setFnPS(lc.shouldComponentUpdate, "shouldComponentUpdate")
+
+      lc.componentWillReceiveProps.foreach { f =>
+        val g = (t: ComponentScopeM[P, S, B], p: WrapObj[P]) => f(t, p.v)
+        spec.updateDynamic("componentWillReceiveProps")(g: ThisFunction)
+      }
+
+      spec.asInstanceOf[ComponentSpec[P, S, B]]
     }
+
+    def create: C = cc(React.createClass(buildSpec))
   }
 }